Added some code
Recoded the slideshow for the website's background and added a hidden tooltip that shows some text when hovered over by your mouse. If you can't find it, it says 'You may have noticed that the Slideshow used may flicker each time a new image is shown, this stops after the first loop of images.' Can't seem to find a fix for that yet.. Also tweaked some text to make the website look better. :+1: